Home Feed state of home is Done

+++ app/src/main/java/iti/android/wheatherappjetpackcompose/dataLayer/source/dto/Current.kt
@@ -1,19 +1,19 @@
 package iti.android.wheatherappjetpackcompose.dataLayer.source.dto
 
 data class Current(
-    var clouds: Int? = null,
+    var clouds: Double? = null,
     val dew_point: Double? = null,
-    val dt: Int? = null,
+    val dt: Double? = null,
     val feels_like: Double? = null,
-    val humidity: Int? = null,
-    val pressure: Int? = null,
-    val sunrise: Int? = null,
-    val sunset: Int? = null,
+    val humidity: Double? = null,
+    val pressure: Double? = null,
+    val sunrise: Double? = null,
+    val sunset: Double? = null,
     val temp: Double? = null,
     val uvi: Double? = null,
-    val visibility: Int? = null,
+    val visibility: Double? = null,
     val weather: List<Weather> = emptyList(),
-    val wind_deg: Int? = null,
+    val wind_deg: Double? = null,
     val wind_gust: Double? = null,
-    val wind_speed: Int? = null,
+    val wind_speed: Double? = null,
 )
